Bugfix: coordinates specification in AR models with unknown rho
If specified with the full argument name `coordinates`, the line used to avoid environment problems duplicated this argument into another called `coord`.
But if the full name is used in this line, the problem arise when using `coord`.

These lines were a workaround to avoid scoping problems when using tests (no problem when datasets were defined in the .GlobalEnv).
However, the correct way to go is to specify the calling environment as an argument of eval.rho().

This grid search process should be externalised for more exhaustive unit testing.

Thanks to K. Vander for reporting.